logistic: refactor file locations
