+++ app/src/main/java/com/example/fridgerec/fragments/InventoryFragment.java
@@ -1,4 +1,4 @@
-package com.example.fridgerec.activities.fragments;
+package com.example.fridgerec.fragments;
 
 import android.os.Bundle;
