+++ app/src/main/java/com/example/fridgerec/fragments/ShoppingCreationFragment.java
@@ -1,4 +1,4 @@
-package com.example.fridgerec.activities.fragments;
+package com.example.fridgerec.fragments;
 
 import android.os.Bundle;
 
@@ -11,56 +11,42 @@ import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
-import android.util.Log;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.Toast;
 
 import com.example.fridgerec.R;
-import com.parse.ParseUser;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link SettingsFragment} factory method to
+ * Use the {@link ShoppingCreationFragment} factory method to
  * create an instance of this fragment.
  */
-public class SettingsFragment extends Fragment {
-  public static final String TAG = "SettingsFragment";
+public class ShoppingCreationFragment extends Fragment {
+  private Toolbar toolbar;
   private NavController navController;
   private AppBarConfiguration appBarConfiguration;
-  private Toolbar toolbar;
-  private Button btnLogout;
 
-  public SettingsFragment(){ }
+  public ShoppingCreationFragment() {
+    // Required empty public constructor
+  }
 
   @Override
   public View onCreateView(LayoutInflater inflater, ViewGroup container,
                            Bundle savedInstanceState) {
-    setHasOptionsMenu(true);
-    return inflater.inflate(R.layout.fragment_settings, container, false);
+    return inflater.inflate(R.layout.fragment_shopping_creation, container, false);
   }
 
   @Override
   public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-    btnLogout = view.findViewById(R.id.btnLogout);
-    toolbar = view.findViewById(R.id.toolbar);
-
+    super.onViewCreated(view, savedInstanceState);
     navController = Navigation.findNavController(view);
 
-    setupToolbar();
+    toolbar = view.findViewById(R.id.toolbar);
 
-    btnLogout.setOnClickListener(new View.OnClickListener() {
-      @Override
-      public void onClick(View v) {
-        logout();
-        navController.navigate(R.id.action_settingsFragment_to_loginFragment);
-      }
-    });
+    setupToolbar();
+    onClickToolbarItem();
+
   }
 
   private void setupToolbar() {
@@ -68,16 +54,12 @@ public class SettingsFragment extends Fragment {
     NavigationUI.setupWithNavController(toolbar, navController, appBarConfiguration);
   }
 
-  private void logout() {
-    ParseUser.logOut();
-    Toast.makeText(getActivity(), "logout success", Toast.LENGTH_SHORT).show();
-  }
-
   private void onClickToolbarItem() {
     toolbar.setOnMenuItemClickListener(item -> {
       switch (item.getItemId()) {
         case R.id.miSave:
-          //TODO: save preferences
+          //TODO: save item
+          navController.navigate(R.id.action_shoppingCreationFragment_to_shoppingFragment);
           return true;
         default:
           return false;
