+++ app/src/main/res/layout/fragment_inventory.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.fragments.InventoryFragment"
+    tools:context=".fragments.InventoryFragment"
     android:orientation="vertical">
 
     <com.google.android.material.appbar.AppBarLayout
