+++ app/src/main/res/layout/fragment_login.xml
@@ -7,7 +7,7 @@
     android:layout_height="wrap_content"
     android:layout_margin="25dp"
     android:orientation="vertical"
-    tools:context=".activities.fragments.LoginFragment"
+    tools:context=".fragments.LoginFragment"
     android:layout_gravity="center_vertical">
 
     <!--    for app logo-->
