+++ app/src/main/res/layout/fragment_settings.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".activities.fragments.SettingsFragment"
+    tools:context=".fragments.SettingsFragment"
     android:orientation="vertical">
 
     <com.google.android.material.appbar.AppBarLayout
